mm-video: vidc: Remove hardcoding of BufCountActual
When metabuffer mode is set, we're harcoding the nBufCountActual to 4.
In most cases we get away with it, except in cases where nBufCountMin
happens to be higher.  This triggers quite a few sanity-check conditions
in the component and driver fail.  Hence to avoid this situation, just
reset nBufCountActual to nBufCountMin.

Change-Id: I4d296574667640bb678bbc4c5cf8b84922ffb6c0
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>